refactor(workspace): :truck: moved `generateMember` file into utils folder
deleted the now-empty 'generators' folder
