+++ tests/fixtures/membersFixtures.ts
@@ -1,4 +1,4 @@
-import generateMember from "../../src/generators/generateMember";
+import generateMember from "../../src/utils/generateMember";
 
 export const averageMembersFixture = ["Vincent", "Archie", "Sen", "Seath"].map(
 	generateMember
